Add rent column

+++ db/migrate/20221107143811_create_tenants.rb
@@ -4,7 +4,7 @@ class CreateTenants < ActiveRecord::Migration[6.1]
       t.string :name
       t.string :email
       t.integer :phone_number
-      t.float :rent
+      t.integer :rent
       t.integer :property_id
       t.timestamps
     end
